+++ elsa/template.h
@@ -25,8 +25,8 @@ public:
   virtual Tag getTag() const { return T_TYPEVAR; }
   virtual sm::string toCString() const;
   virtual sm::string toMLString() const;
-  virtual void sizeInfoInBytes(ellcc::TargetInfo& TI, int &size, int &align) const;
-  virtual void sizeInfoInBits(ellcc::TargetInfo& TI, int &size, int &align) const;
+  virtual void sizeInfoInBytes(ellcc::TargetInfo& TI, uint64_t &size, uint64_t &align) const;
+  virtual void sizeInfoInBits(ellcc::TargetInfo& TI, uint64_t &size, uint64_t &align) const;
   virtual void traverse(TypeVisitor &vis);
 
   // true if this template parameter has been associated with
@@ -61,8 +61,8 @@ public:      // funcs
   virtual Tag getTag() const { return T_PSEUDOINSTANTIATION; }
   virtual sm::string toCString() const;
   virtual sm::string toMLString() const;
-  virtual void sizeInfoInBytes(ellcc::TargetInfo& TI, int &size, int &align) const;
-  virtual void sizeInfoInBits(ellcc::TargetInfo& TI, int &size, int &align) const;
+  virtual void sizeInfoInBytes(ellcc::TargetInfo& TI, uint64_t &size, uint64_t &align) const;
+  virtual void sizeInfoInBits(ellcc::TargetInfo& TI, uint64_t &size, uint64_t &align) const;
   virtual void traverse(TypeVisitor &vis);
 };
 
@@ -92,8 +92,8 @@ public:      // data
   virtual Tag getTag() const { return T_DEPENDENTQTYPE; }
   virtual sm::string toCString() const;
   virtual sm::string toMLString() const;
-  virtual void sizeInfoInBytes(ellcc::TargetInfo& TI, int &size, int &align) const;
-  virtual void sizeInfoInBits(ellcc::TargetInfo& TI, int &size, int &align) const;
+  virtual void sizeInfoInBytes(ellcc::TargetInfo& TI, uint64_t &size, uint64_t &align) const;
+  virtual void sizeInfoInBits(ellcc::TargetInfo& TI, uint64_t &size, uint64_t &align) const;
   virtual void traverse(TypeVisitor &vis);
 };
 
@@ -137,8 +137,8 @@ public:      // funcs
   virtual Tag getTag() const { return T_DEPENDENTSIZEDARRAY; }
   unsigned innerHashValue() const;
   virtual sm::string toMLString() const;
-  virtual void sizeInfoInBytes(ellcc::TargetInfo& TI, int &size, int &align) const;
-  virtual void sizeInfoInBits(ellcc::TargetInfo& TI, int &size, int &align) const;
+  virtual void sizeInfoInBytes(ellcc::TargetInfo& TI, uint64_t &size, uint64_t &align) const;
+  virtual void sizeInfoInBits(ellcc::TargetInfo& TI, uint64_t &size, uint64_t &align) const;
 };      
 
 
@@ -680,8 +680,8 @@ public:      // funcs
   virtual Tag getTag() const { return T_TEMPLATETYPEVAR; }
   virtual sm::string toCString() const;
   virtual sm::string toMLString() const;
-  virtual void sizeInfoInBytes(ellcc::TargetInfo& TI, int &size, int &align) const;
-  virtual void sizeInfoInBits(ellcc::TargetInfo& TI, int &size, int &align) const;
+  virtual void sizeInfoInBytes(ellcc::TargetInfo& TI, uint64_t &size, uint64_t &align) const;
+  virtual void sizeInfoInBits(ellcc::TargetInfo& TI, uint64_t &size, uint64_t &align) const;
   virtual void traverse(TypeVisitor &vis);
 
   // True if this template parameter has been associated with a
